Moved category filter to Heaser.js
